refactor(dashboard): deduplicate store normalizers into store-normalizers.ts
Move normalizeAttentionItem and normalizeRecommendedAction from
room-truth-store.ts and mission-store.ts into store-normalizers.ts.
Remove local normalizeExecutionSummary and normalizeExecutionQueueItem
from room-truth-store.ts (already exported from store-normalizers.ts).
